Refactor additional packages
