+++ src/main/java/duke/Storage.java
@@ -1,5 +1,7 @@
 package duke;
 
+import duke.task.Task;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
