+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,25 +1,35 @@
-package duke;
+package duke.command;
 
-public class SnoozeCommand implements ICommand {
+import duke.*;
+import duke.io.ResponseManager;
+import duke.task.Task;
+import duke.task.TaskManager;
 
-    private final String input;
+/**
+ * This class encapsulates the command to delete a task.
+ * It is triggered by the parser and uses the TaskManager, Ui, and Storage.
+ */
+public class DeleteCommand implements ICommand {
+
+    private String taskIndex;
     private TaskManager tm;
     private ResponseManager responseManager;
     private Storage storage;
-    private Task snoozedTask;
+    private Task deletedTask;
     private String reply;
+
 
     /**
      * Constructor for the command.
      *
-     * @param input The user's input which triggered the creation of this command.
+     * @param taskIndex The index of the task which the user wants deleted.
      */
-    public SnoozeCommand(String input) {
-        this.input = input;
+    public DeleteCommand(String taskIndex) {
+        this.taskIndex = taskIndex;
     }
 
     /**
-     * Changes the deadline/event time.
+     * Deletes the desired task by interacting with the relevant instances as mentioned above.
      *
      * @param tm The TaskManager object controlling the tasks in Duke.
      * @param responseManager The Ui object managing Duke's user interface.
@@ -27,15 +37,15 @@ public class SnoozeCommand implements ICommand {
      */
     public void execute(TaskManager tm, ResponseManager responseManager, Storage storage) {
         try {
-            snoozedTask = tm.snoozeTask(input);
+            deletedTask = tm.deleteTask(taskIndex);
             this.responseManager = responseManager;
-            if (snoozedTask == null) {
+            if (deletedTask == null) {
                 reply = responseManager.getInvalidIndexMessage();
             } else {
-                reply = responseManager.getSnoozeTaskMessage(snoozedTask);
+                reply = responseManager.getTaskDeletionMessage(deletedTask, tm.getTasks().size());
                 storage.updateSave(tm.getTasks());
             }
-        } catch (DukeException.InvalidInputException | DukeException.NoTimeSpecifiedException | DukeException.UnsnoozeableTaskException e) {
+        } catch (DukeException.InvalidInputException e) {
             reply = responseManager.getErrorMessage(e);
         }
     }
