+++ src/main/java/duke/command/DoneCommand.java
@@ -1,30 +1,34 @@
-package duke;
+package duke.command;
+
+import duke.*;
+import duke.io.ResponseManager;
+import duke.task.Task;
+import duke.task.TaskManager;
 
 /**
- * This class encapsulates the command to delete a task.
+ * This class encapsulates the command to complete a task.
  * It is triggered by the parser and uses the TaskManager, Ui, and Storage.
  */
-public class DeleteCommand implements ICommand {
+public class DoneCommand implements ICommand {
 
     private String taskIndex;
     private TaskManager tm;
     private ResponseManager responseManager;
     private Storage storage;
-    private Task deletedTask;
+    private Task completedTask;
     private String reply;
-
 
     /**
      * Constructor for the command.
      *
-     * @param taskIndex The index of the task which the user wants deleted.
+     * @param taskIndex The index of the task which the user wants to mark completed.
      */
-    public DeleteCommand(String taskIndex) {
+    public DoneCommand(String taskIndex) {
         this.taskIndex = taskIndex;
     }
 
     /**
-     * Deletes the desired task by interacting with the relevant instances as mentioned above.
+     * Marks the desired task completed by interacting with the relevant instances as mentioned above.
      *
      * @param tm The TaskManager object controlling the tasks in Duke.
      * @param responseManager The Ui object managing Duke's user interface.
@@ -32,12 +36,14 @@ public class DeleteCommand implements ICommand {
      */
     public void execute(TaskManager tm, ResponseManager responseManager, Storage storage) {
         try {
-            deletedTask = tm.deleteTask(taskIndex);
+            completedTask = tm.completeTask(taskIndex);
             this.responseManager = responseManager;
-            if (deletedTask == null) {
+            this.tm = tm;
+            this.storage = storage;
+            if (completedTask == null) {
                 reply = responseManager.getInvalidIndexMessage();
             } else {
-                reply = responseManager.getTaskDeletionMessage(deletedTask, tm.getTasks().size());
+                reply = responseManager.getTaskCompletionMessage(completedTask, tm.getTasks().size());
                 storage.updateSave(tm.getTasks());
             }
         } catch (DukeException.InvalidInputException e) {
