+++ src/main/java/duke/command/ICommand.java
@@ -1,4 +1,8 @@
-package duke;
+package duke.command;
+
+import duke.io.ResponseManager;
+import duke.Storage;
+import duke.task.TaskManager;
 
 /**
  * This interface is used by all commands and ensures that commands implement the execute() method.
