+++ src/main/java/duke/io/HelpPopUp.java
@@ -1,11 +1,9 @@
-package duke;
+package duke.io;
 
 import java.io.IOException;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.control.Label;
-import javafx.scene.control.TextArea;
-import javafx.scene.control.TextField;
 import javafx.scene.layout.HBox;
 
 /**
