+++ src/main/java/duke/io/UserDialogBox.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.io;
 
 import java.io.IOException;
 import java.util.Collections;
